requirements.txt
dhooks
python_socketio
eventlet